Imported upstream version '1.8.0' of 'upstream'

+++ planning/behavior_velocity_planner/autoware_behavior_velocity_planner/src/node.cpp
@@ -18,7 +18,6 @@
 #include <autoware/motion_utils/trajectory/path_with_lane_id.hpp>
 #include <autoware/motion_utils/trajectory/trajectory.hpp>
 #include <autoware/velocity_smoother/smoother/analytical_jerk_constrained_smoother/analytical_jerk_constrained_smoother.hpp>
-#include <autoware_lanelet2_extension/utility/message_conversion.hpp>
 #include <autoware_utils_pcl/transforms.hpp>
 #include <autoware_utils_rclcpp/parameter.hpp>
 #include <tf2_eigen/tf2_eigen.hpp>
@@ -405,7 +404,7 @@ void BehaviorVelocityPlannerNode::publishDebugMarker(const autoware_planning_msg
   for (size_t i = 0; i < path.points.size(); ++i) {
     visualization_msgs::msg::Marker marker;
     marker.header = path.header;
-    marker.id = i;
+    marker.id = static_cast<int>(i);
     marker.type = visualization_msgs::msg::Marker::ARROW;
     marker.pose = path.points.at(i).pose;
     marker.scale.y = marker.scale.z = 0.05;
